Added new files and changed the setup of where the Profile files are stored
