--- src/pages/Profile.module.css
@@ -1,30 +1,0 @@
-.reserved_missions {
-  width: 100%;
-  max-width: 800px;
-  margin: 0 auto;
-}
-
-.missions_table {
-  width: 50%;
-  border-collapse: collapse;
-  margin-top: 20px;
-}
-
-.missions_table th,
-.missions_table td {
-  border: 1px solid #ccc;
-  padding: 8px;
-  text-align: left;
-}
-
-.missions_table th {
-  background-color: #f2f2f2;
-}
-
-.missions_table tr:nth-child(even) {
-  background-color: #f2f2f2;
-}
-
-.missions_table tr:hover {
-  background-color: #ddd;
-}
